+++ Swith-Server/src/main/java/com/swith/api/attendance/dto/GetUserAttendanceRes.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.response;
+package com.swith.api.attendance.dto;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
