+++ Swith-Server/src/main/java/com/swith/api/attendance/dto/UserAttendanceInfo.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.response;
+package com.swith.api.attendance.dto;
 
 public interface UserAttendanceInfo {
     Long getUserIdx();
