--- Swith-Server/src/main/java/com/swith/api/dto/ErrorResponse.java
@@ -1,13 +1,0 @@
-package com.swith.api.dto;
-
-public class ErrorResponse {
-    private String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
