+++ Swith-Server/src/main/java/com/swith/api/groupinfo/controller/GroupInfoController.java
@@ -9,9 +9,9 @@ import com.swith.api.groupinfo.dto.PostGroupInfoRes;
 import com.swith.api.groupinfo.dto.GetGroupInfoSearchReq;
 import com.swith.api.groupinfo.dto.PatchEndGroupReq;
 import com.swith.api.groupinfo.dto.PatchGroupInfoReq;
-import com.swith.api.dto.response.GetEachGroupInfoRes;
-import com.swith.api.dto.response.GetGroupInfoSearchRes;
-import com.swith.api.dto.response.GetHomeGroupInfoRes;
+import com.swith.api.groupinfo.dto.GetEachGroupInfoRes;
+import com.swith.api.groupinfo.dto.GetGroupInfoSearchRes;
+import com.swith.api.groupinfo.dto.GetHomeGroupInfoRes;
 import com.swith.domain.user.service.UserService;
 import com.swith.global.error.exception.BaseException;
 import com.swith.domain.groupinfo.service.GroupInfoService;
