+++ Swith-Server/src/main/java/com/swith/api/rating/dto/PostRatingReq.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.request;
+package com.swith.api.rating.dto;
 
 
 import lombok.AllArgsConstructor;
