+++ Swith-Server/src/main/java/com/swith/api/rating/dto/PostRatingRes.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.request;
+package com.swith.api.rating.dto;
 
 
 import lombok.AllArgsConstructor;
@@ -6,15 +6,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class PostRatingStarReq {
-
-    private Long raterIdx; //평가자
-    private List<Start> star;
-
+public class PostRatingRes {
+    private Long userIdx;
+    private String nickname;
 }
