+++ Swith-Server/src/main/java/com/swith/api/rating/dto/PostRatingStarReq.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.response;
+package com.swith.api.rating.dto;
 
 
 import lombok.AllArgsConstructor;
@@ -6,11 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class PostRatingRes {
-    private Long userIdx;
-    private String nickname;
+public class PostRatingStarReq {
+
+    private Long raterIdx; //평가자
+    private List<Start> star;
+
 }
