+++ Swith-Server/src/main/java/com/swith/api/rating/dto/PostRatingStarRes.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.response;
+package com.swith.api.rating.dto;
 
 
 import lombok.Data;
