+++ Swith-Server/src/main/java/com/swith/api/session/controller/SessionController.java
@@ -5,7 +5,7 @@ import com.swith.api.common.dto.BaseResponse;
 import com.swith.api.common.constant.BaseResponseStatus;
 import com.swith.api.dto.request.PatchSessionReq;
 import com.swith.api.dto.response.GetGroupInfoRes;
-import com.swith.api.dto.request.PostSessionReq;
+import com.swith.api.session.dto.PostSessionReq;
 import com.swith.api.dto.response.GetSessionTabRes;
 import com.swith.domain.user.service.UserService;
 import com.swith.domain.session.service.SessionService;
