+++ Swith-Server/src/main/java/com/swith/api/session/dto/PostSessionReq.java
@@ -1,4 +1,4 @@
-package com.swith.api.dto.request;
+package com.swith.api.session.dto;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.*;
