+++ Swith-Server/src/main/java/com/swith/domain/groupinfo/repository/GroupInfoRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.swith.domain.groupinfo.repository;
 
 import com.swith.api.groupinfo.dto.GetGroupInfoSearchReq;
-import com.swith.api.dto.response.GetGroupInfoSearchRes;
+import com.swith.api.groupinfo.dto.GetGroupInfoSearchRes;
 import com.querydsl.jpa.impl.JPAQuery;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
