+++ Swith-Server/src/main/java/com/swith/domain/groupinfo/service/GroupInfoService.java
@@ -18,12 +18,12 @@ import com.swith.domain.session.repository.SessionRepository;
 import com.swith.global.error.exception.BaseException;
 import com.swith.api.common.constant.BaseResponseStatus;
 import com.swith.api.groupinfo.dto.PatchGroupInfoReq;
-import com.swith.api.dto.response.GetEachGroupInfoRes;
-import com.swith.api.dto.response.GetHomeGroupInfoRes;
+import com.swith.api.groupinfo.dto.GetEachGroupInfoRes;
+import com.swith.api.groupinfo.dto.GetHomeGroupInfoRes;
 import com.swith.api.groupinfo.dto.PostGroupInfoReq;
 import com.swith.api.groupinfo.dto.PostGroupInfoRes;
 import com.swith.api.groupinfo.dto.GetGroupInfoSearchReq;
-import com.swith.api.dto.response.GetGroupInfoSearchRes;
+import com.swith.api.groupinfo.dto.GetGroupInfoSearchRes;
 import com.swith.external.firebase.FirebaseCloudMessageService;
 import com.querydsl.jpa.impl.JPAQuery;
 import org.springframework.beans.factory.annotation.Autowired;
